[Improvement] Refactor FlinkProcessRegistrar to be more extensible
